Move HeaderDayDate composable into "designsystem" package.
